Apply more tweaks to the way webpack config package is structured
