+++ webpack.config.js
@@ -13,7 +13,7 @@ const { basename } = require( 'path' );
  */
 const CustomTemplatedPathPlugin = require( '@wordpress/custom-templated-path-webpack-plugin' );
 const LibraryExportDefaultPlugin = require( '@wordpress/library-export-default-webpack-plugin' );
-const webpackConfig = require( '@wordpress/webpack-config' );
+const { camelCaseDash, config } = require( '@wordpress/webpack-config' );
 
 /**
  * Internal dependencies
@@ -22,38 +22,28 @@ const { dependencies } = require( './package' );
 
 const WORDPRESS_NAMESPACE = '@wordpress/';
 
-/**
- * Given a string, returns a new string with dash separators converted to
- * camelCase equivalent. This is not as aggressive as `_.camelCase` in
- * converting to uppercase, where Lodash will also capitalize letters
- * following numbers.
- *
- * @param {string} string Input dash-delimited string.
- *
- * @return {string} Camel-cased string.
- */
-function camelCaseDash( string ) {
-	return string.replace(
-		/-([a-z])/g,
-		( match, letter ) => letter.toUpperCase()
-	);
-}
-
 const gutenbergPackages = Object.keys( dependencies )
 	.filter( ( packageName ) => packageName.startsWith( WORDPRESS_NAMESPACE ) )
 	.map( ( packageName ) => packageName.replace( WORDPRESS_NAMESPACE, '' ) );
 
-webpackConfig.entry = gutenbergPackages.reduce( ( memo, packageName ) => {
+config.entry = gutenbergPackages.reduce( ( memo, packageName ) => {
 	const name = camelCaseDash( packageName );
 	memo[ name ] = `./packages/${ packageName }`;
 	return memo;
 }, {} );
 
-webpackConfig.plugins.push(
+config.output = {
+	filename: './build/[basename]/index.js',
+	path: __dirname,
+	library: [ 'wp', '[name]' ],
+	libraryTarget: 'this',
+};
+
+config.plugins.push(
 	// Create RTL files with a -rtl suffix
 	new WebpackRTLPlugin( {
 		suffix: '-rtl',
-		minify: process.env.NODE_ENV === 'production' ? { safe: true } : false,
+		minify: config.mode === 'production' ? { safe: true } : false,
 	} ),
 	new CustomTemplatedPathPlugin( {
 		basename( path, data ) {
@@ -90,7 +80,7 @@ webpackConfig.plugins.push(
 			to: `./build/${ packageName }/`,
 			flatten: true,
 			transform: ( content ) => {
-				if ( webpackConfig.mode === 'production' ) {
+				if ( config.mode === 'production' ) {
 					return postcss( [
 						require( 'cssnano' )( {
 							preset: [ 'default', {
@@ -109,4 +99,4 @@ webpackConfig.plugins.push(
 	)
 );
 
-module.exports = webpackConfig;
+module.exports = config;
